second deployment with service list
Changed service import to defer the delete to a background thread
and to add services to the data store in batches of 10 items. A
GAE channel is provided that allows a client to watch the progress
of an import.  Added page to create a new service. Added a page to
view authorised devices and APIs. Now using wtforms to handle forms
and mimeparse to handle content type negotiation.